refactor(transit-display): extract TransitDisplaysContainer
Move TransitDisplaysContainer into its own file so the presentational
TransitDisplays can be swapped without touching the orchestration. The
container keeps the shared concerns (build the dataset via
buildTransitDisplayDataSet, scroll-fade scaffold, callback wiring) and
depends only on the TransitDisplaysProps contract, so an alternate
signage style is a drop-in replacement that reuses the same data.

Also thread infoLevel through the display components (verbose gates the
mode emojis) and rename the entry's row prop to data, updating the
TransitDisplayEntry stories to match.
